+++ doc/CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [master]
+### Added
+- InformationCollectionMapper, InformationCollectionStruct and InformationCollectionType
+
 ## [1.0.3] - 2017-04-21
 ### Added
 - Symfony 3.x compatibility